extmod/ujson: Add support for CPython-compatible nan/inf handling.
Modify the JSON loading code so it can parse the JavaScript floating
point representations NaN/Infinity/-Infinity, and support the allow_nan
keyword argument for dump() and dumps() which controls whether those
representations get created or else lead to raising a ValueError.
All of this is the same as CPython's implementation and when enabled
fixes the inconsistency in the current code which unconditionally
creates lowercase nan/inf strings but doesn't parse them.