py: Make import recognize and load native modules (.pyd or .so files).
Introduce CPython-style dynamic module loading: upon import, if no .py
file is found, check if a .pyd or .so file exists on the module search
path and if so load it using the dynamic loader, lookup the symbol
'init_<modulename>' and call it as a function which returns the actual
uPy module.  Implemented for unix and windows ports.

Signed-off-by: stijn <stijn@ignitron.net>