windows: Use one single job for Appveyor builds.
Run all builds and tests for all configurations and platforms from
a single script.  This doesn't change what gets built, but in
comparison with using multiple jobs this has the advantage all build
results are in one place which requires less clicking in the Appveyor
UI to get to those results, and the builds finish slightly faster
because only one job needs to be started.  Additional advantage is
the entire script can be used to build locally as well, that just
requires setting the MicroPythonDir, Bash and MICROPY_CPYTHON3 vars.