windows: Build separate micropython dll and executable.
To enable building native dll modules export commonly used core
functions and data; functions are exported with a .def file so no code
change is required but exporting data using .def file yields multiple
copies of data so a dllexport/import macro seems the only solution.

MicroPython itself is split in a 2 projects: the original is copied
to micropythoncore.vcxproj and altered to produce a .dll and corresponding
import .lib, and the new micropython.vcxproj is just a stub around main().

Also provide a property sheet with default values for external modules,
and run.h/run.c with extra functions for embedding the interpreter.

Signed-off-by: stijn <stijn@ignitron.net>